+++ src/baobab_mtg_deckbuilder/optimization/__init__.py
@@ -1,0 +1,33 @@
+"""Optimisation itérative de decks (stratégies de recherche locales et faisceau)."""
+
+from baobab_mtg_deckbuilder.optimization.beam_search_optimization_strategy import (
+    BeamSearchOptimizationStrategy,
+)
+from baobab_mtg_deckbuilder.optimization.deck_optimization_iteration import (
+    DeckOptimizationIteration,
+)
+from baobab_mtg_deckbuilder.optimization.deck_optimization_request import DeckOptimizationRequest
+from baobab_mtg_deckbuilder.optimization.deck_optimization_result import DeckOptimizationResult
+from baobab_mtg_deckbuilder.optimization.deck_optimization_strategy import DeckOptimizationStrategy
+from baobab_mtg_deckbuilder.optimization.deck_search_state import DeckSearchState
+from baobab_mtg_deckbuilder.optimization.hill_climbing_optimization_strategy import (
+    HillClimbingOptimizationStrategy,
+)
+from baobab_mtg_deckbuilder.optimization.iterative_improvement_strategy import (
+    IterativeImprovementStrategy,
+)
+from baobab_mtg_deckbuilder.optimization.optimization_evaluation import (
+    default_optimization_evaluation,
+)
+
+__all__ = [
+    "BeamSearchOptimizationStrategy",
+    "DeckOptimizationIteration",
+    "DeckOptimizationRequest",
+    "DeckOptimizationResult",
+    "DeckOptimizationStrategy",
+    "DeckSearchState",
+    "HillClimbingOptimizationStrategy",
+    "IterativeImprovementStrategy",
+    "default_optimization_evaluation",
+]